rename bytecode.* to instruction.*
