--- src/bytecodevm/bytecode.cpp
@@ -1,1 +1,0 @@
-#include "bytecode.hpp"
